* Add comments for #endif / #else * Speed up compilation * regression in GetResourceLimit: returns UINT_MAX (-> -1) instead of INT_MAX * fixed some more compiler warnings * removed unnecessary 'package require' * Merged Patrick Mattha:ei's fixes for PATH_MAX on GNU hurd * Moved third-party code into third-party directory * Removed unnecessary 'package require' statements in scripts
git-svn-id: http://svn.beutner.name/trunk/sbnc@1086 b3a6972a-4407-0410-a3f0-9bc6c1699d69

+++ src/Nick.cpp
@@ -56,7 +56,7 @@ CNick::~CNick() {
 	free(m_Realname);
 	free(m_Server);
 
-	for (unsigned int i = 0; i < m_Tags.GetLength(); i++) {
+	for (int i = 0; i < m_Tags.GetLength(); i++) {
 		free(m_Tags[i].Name);
 		free(m_Tags[i].Value);
 	}
@@ -184,7 +184,7 @@ bool CNick::RemovePrefix(char Prefix) {
 		return false;
 	}
 
-	for (unsigned int i = 0; i < LengthPrefixes; i++) {
+	for (size_t i = 0; i < LengthPrefixes; i++) {
 		if (m_Prefixes[i] != Prefix) {
 			Copy[a++] = m_Prefixes[i];
 		}
@@ -432,7 +432,7 @@ bool CNick::SetIdleSince(time_t Time) {
  * @param Name the name of the tag
  */
 const char *CNick::GetTag(const char *Name) const {
-	for (unsigned int i = 0; i < m_Tags.GetLength(); i++) {
+	for (int i = 0; i < m_Tags.GetLength(); i++) {
 		if (strcasecmp(m_Tags[i].Name, Name) == 0) {
 			return m_Tags[i].Value;
 		}
@@ -456,7 +456,7 @@ bool CNick::SetTag(const char *Name, const char *Value) {
 		return false;
 	}
 
-	for (unsigned int i = 0; i < m_Tags.GetLength(); i++) {
+	for (int i = 0; i < m_Tags.GetLength(); i++) {
 		if (strcasecmp(m_Tags[i].Name, Name) == 0) {
 			free(m_Tags[i].Name);
 			free(m_Tags[i].Value);
